pyproject.toml
[tool.poetry]
name = "fdk-reports-bff"
version = "1.0.0"
description = ""
authors = ["NilsOveTen <nils.ove.tendenes@digdir.no>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "2.0.1"
flask-cors = "3.0.10"
flask-restful = "0.3.9"
python-dotenv = "0.19.0"
requests = "2.26.0"
gunicorn = "20.1.0"
asyncstdlib = "3.9.2"
httpx = "0.18.2"
elasticsearch = "7.12.1"
apscheduler = "3.7.0"
python-dateutil = "2.8.2"
python-json-logger = "2.0.2"

[tool.poetry.dev-dependencies]
pytest = "6.2.4"
pytest-mock = "3.6.1"
mock = "4.0.3"
isort = "5.9.3"
black = "^21.5b1"
flake8 = "^3.9.2"
pep8-naming = "^0.11.1"
safety = "^1.10.3"
flake8-import-order = "^0.18.1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.4.3"
flake8-annotations = "^2.6.2"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["fdk_fulltext_search"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
